Move map into views/components
